+++ lib/commands/irmin_core.ml
@@ -214,6 +214,8 @@ module Key_ :
     val key_to_string : t -> string
     val key_to_path : t -> string
     val view_key_to_path : t -> string
+    val key_to_workpath : t -> string
+    val view_key_to_workpath : t -> string
     val assert_key : t -> unit
   end with type t = View.key =
   struct
@@ -290,6 +292,10 @@ module Key_ :
      in
      acc
 
+    (* convert key to git workdir path, keep all components *)
+    let key_to_workpath key = 
+      String.concat "/" key
+
     (* convert view key (relative key) to path, remove "mailboxes" *)
     let view_key_to_path key = 
      let (_,acc) =
@@ -306,6 +312,10 @@ module Key_ :
      in
      acc
 
+    (* convert view key (relative key) to workdir path, keep all components *)
+    let view_key_to_workpath key = 
+      String.concat "/" key
+
   end
 
 let get_irmin_path user config =
@@ -313,19 +323,93 @@ let get_irmin_path user config =
   | None -> config.irmin_path
   | Some user -> Regex.replace ~regx:"%user%" ~tmpl:user config.irmin_path
 
-module IrminIntf :
+module type GitIntf =
   sig
     type store
+    type view
     val create : ?user:string -> Server_config.imapConfig -> store Lwt.t
     val remove : store -> Key_.t -> unit Lwt.t
     val read_exn : store -> Key_.t -> string Lwt.t
     val mem : store -> Key_.t -> bool Lwt.t
-    val list : store -> Key_.t -> View.key list Lwt.t
-    val update_view : store -> Key_.t -> View.t -> unit Lwt.t
-    val read_view : store -> Key_.t -> View.t Lwt.t
-  end =
+    val list : store -> Key_.t -> Key_.t list Lwt.t
+    val update : store -> Key_.t -> string -> unit Lwt.t
+    val update_view : store -> Key_.t -> view -> unit Lwt.t
+    val read_view : store -> Key_.t -> view Lwt.t
+  end
+
+(* workdir creates and updates files in git working directory
+ * the files (metadata, indeces, not messages) are overwritten! this is conceptually implementation of maildir
+ *)
+module GitWorkdirIntf : GitIntf with type store = string 
+  and type view = string =
+  struct
+    type store = string
+    type view = string (* reference to path *)
+
+    let get_path store key =
+      Filename.concat store (Key_.key_to_workpath key)
+
+    let create ?user config =
+      return (get_irmin_path user config)
+
+    let remove store key =
+      let path = get_path store key in
+      exists path S_REG >>= fun res ->
+      if res then
+        Lwt_unix.unlink path
+      else (
+        Lwt_unix.system ("rm -rf " ^ path) >>= fun _ ->
+        return ()
+      )
+
+    let read_exn store key =
+      let path = get_path store key in
+      with_file ~lock:true path ~flags:[O_RDONLY] ~perms:0o666
+      ~mode:Lwt_io.Input ~f:(fun ch ->
+        Lwt_io.read ch
+      )
+
+    let mem store key =
+      let path = get_path store key in
+      exists path ~alt:S_DIR S_REG 
+
+    let list store key =
+      let path = get_path store key in
+      files_of_directory path (fun acc file -> 
+        let key = Key_.add_list key [file] in
+        return (key :: acc)
+      ) [] >>= function
+      | `Ok l -> return l
+      | `NoDir -> return [] (* directory doesn't exist -> mailbox with no submailboxes *)
+
+    let update store key data =
+      let path = get_path store key in
+      let base = Regex.replace ~regx:" " ~tmpl:"\\ " (Filename.dirname path) in
+      Lwt_unix.system ("mkdir -p " ^ base) >>= fun _ ->
+      Lwt_unix.system ("chmod 777 " ^ base) >>= fun _ ->
+      with_file ~lock:true path ~flags:[O_CREAT;O_WRONLY] ~perms:0o666
+      ~mode:Lwt_io.Output ~f:(fun ch ->
+        Lwt_io.write ch data
+      )
+
+    (* update_view doesn't do anything in workdir
+     * all updates happen by individual updates in GitIntf_tr
+     *)
+    let update_view store key v =
+      return ()
+
+    (* just return the path to the view *)
+    let read_view store key = 
+      return (Key_.view_key_to_workpath key)
+
+  end
+
+module IrminIntf : GitIntf with type store = (string -> View.db) and 
+  type view = View.t =
   struct
     type store = (string -> View.db)
+    type view = View.t
+
     let fmt t x = Printf.ksprintf (fun str -> t str) x
     let path () = String.concat "/"
 
@@ -356,9 +440,11 @@ module IrminIntf :
       Key_.assert_key key;
       Store.list (fmt store "List the contents of %a" path key) key
 
+    let update store key data =
+      Store.update (fmt store "Update %a." path key) key data
+
     let update_view store key view =
       Key_.assert_key key;
-      (*Printf.printf "------ store update_view %s\n%!" (Key_.key_to_string * key);*)
       let msg =
         let buf = Buffer.create 1024 in
         let path buf key = Buffer.add_string buf (String.concat "/" key) in
@@ -377,13 +463,11 @@ module IrminIntf :
 
     let read_view store key =
       Key_.assert_key key;
-      (*Printf.printf "------ reading view %s\n%!" (Key_.key_to_string key);*)
       View.of_path (fmt store "Reading %a" path key) key
 
   end
 
-
-module IrminIntf_tr :
+module type GitIntf_tr =
   sig
     type transaction
     val remove_view : transaction -> unit Lwt.t
@@ -393,25 +477,107 @@ module IrminIntf_tr :
     val update : transaction -> Key_.t -> string -> unit Lwt.t
     val read : transaction -> Key_.t -> string option Lwt.t
     val read_exn : transaction -> Key_.t -> string Lwt.t
-    val list : transaction -> Key_.t -> View.key list Lwt.t
+    val list : transaction -> Key_.t -> Key_.t list Lwt.t
     val remove : transaction -> Key_.t -> unit Lwt.t
     val mem : transaction -> Key_.t -> bool Lwt.t
-  end =
+  end
+
+module GitWorkdirIntf_tr : GitIntf_tr with type transaction =
+  GitWorkdirIntf.store * string * Key_.t * bool ref =
+  struct 
+    type transaction = GitWorkdirIntf.store * string * Key_.t * bool ref
+
+    let begin_transaction ?user config key =
+      GitWorkdirIntf.create ?user config >>= fun store ->
+      GitWorkdirIntf.read_view store key >>= fun view ->
+      return (store, view, key, ref false)
+
+    let end_transaction tr =
+      let (store,view,key,dirty) = tr in
+      if !dirty = true then (
+        GitWorkdirIntf.update_view store key view >>= fun () ->
+        dirty := false;
+        return ()
+      ) else
+        return ()
+
+    let remove_view tr =
+      let (store,_,key,_) = tr in
+      GitWorkdirIntf.remove store key
+
+    let move_view tr key2 =
+      let (store,view,_,_) = tr in
+      Lwt_unix.rename (Filename.concat store view)
+        (Filename.concat (Key_.key_to_workpath key2) view)
+
+    let add_key_to_view view key =
+      Key_.add_list (Key_.t_of_path view) key
+
+    let update tr key data =
+      Key_.assert_key key;
+      let (store,view,_,dirty) = tr in
+      GitWorkdirIntf.update store (add_key_to_view view key) data >>= fun () ->
+      dirty := true;
+      return ()
+
+    let read_exn tr key =
+      Key_.assert_key key;
+      let (store,view,_,_) = tr in
+      GitWorkdirIntf.read_exn store (add_key_to_view view key)
+
+    let read tr key =
+      catch (fun () -> 
+        read_exn tr key >>= fun res -> 
+        return (Some res)
+      ) (fun _ -> return None)
+
+    let list tr key =
+      Key_.assert_key key;
+      let (store,view,_,_) = tr in
+      (* view starts at imaplet/user *)
+      let path = Key_.key_to_workpath (add_key_to_view view key) in
+      let path = Filename.concat store path in
+      files_of_directory path (fun acc file ->
+        (* include the parent directory (key) relative to the view *)
+        let key = Key_.add_list key [file] in
+        return (key :: acc)
+      ) [] >>= function
+      | `Ok l -> return l
+      | `NoDir -> return []
+
+    let remove tr key =
+      Key_.assert_key key;
+      let (store,view,_,dirty) = tr in
+      GitWorkdirIntf.remove store (add_key_to_view view key) >>= fun res -> 
+      dirty := true;
+      return ()
+
+    let mem tr key =
+      Key_.assert_key key;
+      let (store,view,_,_) = tr in
+      GitWorkdirIntf.mem store (add_key_to_view view key)
+
+    let tr_key tr =
+      let (_,_,key,_) = tr in
+      key
+
+  end
+
+type irmin_trans = IrminIntf.store * View.t * Key_.t * bool ref
+
+module IrminIntf_tr : GitIntf_tr with type transaction = irmin_trans =
   struct
-    type transaction = IrminIntf.store * View.t * Key_.t * bool ref
+    type transaction = irmin_trans
 
     let begin_transaction ?user config key =
       Key_.assert_key key;
       IrminIntf.create ?user config >>= fun store ->
-      (*Printf.printf "------ creating view %s\n%!" (Key_.key_to_string key);*)
       IrminIntf.read_view store key >>= fun view ->
       return (store,view,key,ref false)
 
     let end_transaction tr =
       let (store,view,key,dirty) = tr in
       if !dirty = true then (
-        (*Printf.printf "++++++++++++++++++ commiting %s!!!\n%!"
-        (Key_.key_to_string key);*)
         IrminIntf.update_view store key view >>= fun () ->
         dirty := false;
         return ()
@@ -426,10 +592,8 @@ module IrminIntf_tr :
       let (store,view,_,_) = tr in
       IrminIntf.update_view store key2 view 
 
-
     let update tr key data =
       Key_.assert_key key;
-      (*Printf.printf "------ store view.update %s\n" (Key_.key_to_string * key);*)
       let (_,view,_,dirty) = tr in
       View.update view key data >>= fun () ->
       dirty := true;
@@ -447,13 +611,11 @@ module IrminIntf_tr :
 
     let list tr key =
       Key_.assert_key key;
-      (*Printf.printf "------ store list %s\n%!" (Key_.key_to_string key);*)
       let (_,view,_,_) = tr in
       View.list view key
 
     let remove tr key =
       Key_.assert_key key;
-      (*Printf.printf "------ store remove %s\n" (Key_.key_to_string key);*)
       let (_,view,_,dirty) = tr in
       View.remove_rec view key >>= fun () ->
       dirty := true;
@@ -470,12 +632,7 @@ module IrminIntf_tr :
 
   end
 
-  type mailbox_ = {user:string;mailbox:string;trans:IrminIntf_tr.transaction;
-  index:int list option ref;config: Server_config.imapConfig;mbox_key:
-    Key_.t;pubpriv: Ssl_.keys}
-
-(* consistency TBD *)
-module IrminMailbox :
+module type GitMailboxIntf = 
   sig
     type t
     val create : Server_config.imapConfig -> string -> string -> Ssl_.keys -> t Lwt.t
@@ -504,10 +661,16 @@ module IrminMailbox :
     val uid_to_seq : t -> int -> int option Lwt.t
     val subscribe : t -> unit Lwt.t
     val unsubscribe : t -> unit Lwt.t
-  end with type t = mailbox_ = 
+  end
+
+module GitMailboxMake 
+  (GI:GitIntf) 
+  (GI_tr:GitIntf_tr) : GitMailboxIntf =
   struct 
     (* user * mailbox * is-folder * irmin key including the mailbox *)
-    type t = mailbox_
+    type t = {user:string;mailbox:string;trans:GI_tr.transaction;
+  index:int list option ref;config: Server_config.imapConfig;mbox_key:
+    Key_.t;pubpriv: Ssl_.keys}
 
     let get_key mbox_key = function
     | `Metamailbox -> Key_.add_path mbox_key "meta"
@@ -523,12 +686,14 @@ module IrminMailbox :
      * changes commited to Irmin
      *)
     let create config user path keys =
+      (* view starts at /imaplet/user, then mbox_key continues at
+       * mailboxes/Foo/mailboxes/Foo1 etc *)
       let (mailbox,mbox_key) = Key_.mailbox_of_path path in
-      IrminIntf_tr.begin_transaction ~user config (Key_.create_account user) >>= fun trans ->
+      GI_tr.begin_transaction ~user config (Key_.create_account user) >>= fun trans ->
         return {user;mailbox;trans;index=ref None;config;mbox_key;pubpriv=keys}
 
     let commit mbox =
-      IrminIntf_tr.end_transaction mbox.trans
+      GI_tr.end_transaction mbox.trans
 
     (* create mailbox metadata and index stores *)
     let create_mailbox mbox =
@@ -536,30 +701,29 @@ module IrminMailbox :
       let metadata = empty_mailbox_metadata ~uidvalidity:(new_uidvalidity())()
         ~selectable:true in
       let sexp = sexp_of_mailbox_metadata metadata in
-      IrminIntf_tr.update mbox.trans key (Sexp.to_string sexp) >>= fun () ->
+      GI_tr.update mbox.trans key (Sexp.to_string sexp) >>= fun () ->
       let key = get_key mbox.mbox_key `Index in
       let sexp = sexp_of_list (fun i -> Sexp.of_string i) [] in
-      IrminIntf_tr.update mbox.trans key (Sexp.to_string sexp) 
+      GI_tr.update mbox.trans key (Sexp.to_string sexp) 
 
     let delete_mailbox mbox =
-      IrminIntf_tr.remove_view mbox.trans
+      GI_tr.remove_view mbox.trans
 
     (* how to make this the transaction? TBD *)
     let move_mailbox mbox path =
       let (_,key2) = Key_.mailbox_of_path path in
-      IrminIntf_tr.move_view mbox.trans key2 >>
-      IrminIntf_tr.remove_view mbox.trans
+      GI_tr.move_view mbox.trans key2 >>
+      GI_tr.remove_view mbox.trans
 
     let read_mailbox_metadata mbox =
-      IrminIntf_tr.read_exn mbox.trans (get_key mbox.mbox_key `Metamailbox) >>= fun sexp_str ->
+      GI_tr.read_exn mbox.trans (get_key mbox.mbox_key `Metamailbox) >>= fun sexp_str ->
       let sexp = Sexp.of_string sexp_str in
       return (mailbox_metadata_of_sexp sexp)
 
     let update_mailbox_metadata mbox metadata =
-      (*Printf.printf "updating mailbox metadata %d\n%!" metadata.uidnext;*)
       let sexp = sexp_of_mailbox_metadata metadata in
       let key = get_key mbox.mbox_key `Metamailbox in
-      IrminIntf_tr.update mbox.trans key (Sexp.to_string sexp)
+      GI_tr.update mbox.trans key (Sexp.to_string sexp)
 
     let exists mbox =
       catch (fun () ->
@@ -571,7 +735,7 @@ module IrminMailbox :
     let exists_key mbox key =
       catch (fun () ->
         let key = Key_.add_list mbox.mbox_key (Key_.add_path key "meta") in
-        IrminIntf_tr.read_exn mbox.trans key >>= fun metadata_sexp_str ->
+        GI_tr.read_exn mbox.trans key >>= fun metadata_sexp_str ->
         let metadata_sexp = Sexp.of_string metadata_sexp_str in
         let metadata = mailbox_metadata_of_sexp metadata_sexp in
         return (if metadata.selectable then `Mailbox else `Folder)
@@ -584,7 +748,7 @@ module IrminMailbox :
     let read_index_uid mbox =
       match mbox.!index with
       | None ->
-        IrminIntf_tr.read_exn mbox.trans (get_key mbox.mbox_key `Index) >>= fun index_sexp_str ->
+        GI_tr.read_exn mbox.trans (get_key mbox.mbox_key `Index) >>= fun index_sexp_str ->
         let uids = list_of_sexp 
           (fun i -> int_of_string (Sexp.to_string i))
           (Sexp.of_string index_sexp_str) in
@@ -594,7 +758,7 @@ module IrminMailbox :
 
     let update_index_uids mbox uids = 
       mbox.index := Some uids;
-      IrminIntf_tr.update mbox.trans (get_key mbox.mbox_key `Index) 
+      GI_tr.update mbox.trans (get_key mbox.mbox_key `Index) 
         (Sexp.to_string (sexp_of_list (fun i -> Sexp.of_string (string_of_int i)) uids ))
 
     let update_index_uid mbox uid =
@@ -614,7 +778,7 @@ module IrminMailbox :
       | Some (i,_) -> return (Some ((List.length uids) - i))
 
     let get_hashes mbox uid =
-      IrminIntf_tr.read_exn mbox.trans (get_key mbox.mbox_key (`Hashes uid)) >>= fun h ->
+      GI_tr.read_exn mbox.trans (get_key mbox.mbox_key (`Hashes uid)) >>= fun h ->
       return (hashes_of_sexp (Sexp.of_string h))
 
     let update_hashes mbox uid msg_hash postmark headers content attachments =
@@ -625,17 +789,17 @@ module IrminMailbox :
         content = "2";
         attachments;
       } in
-      IrminIntf_tr.update mbox.trans (get_key mbox.mbox_key (`Hashes uid)) 
+      GI_tr.update mbox.trans (get_key mbox.mbox_key (`Hashes uid)) 
         (Sexp.to_string (sexp_of_hashes h)) >>
       return h
 
     let append_message_raw mbox ~parse message_metadata =
       let update msg_hash contid data =
         let key = get_key mbox.mbox_key (`Storage (mbox.user,msg_hash,contid)) in
-        IrminIntf_tr.update mbox.trans key data
+        GI_tr.update mbox.trans key data
       in
       let uid = message_metadata.uid in
-      IrminIntf_tr.update mbox.trans (get_key mbox.mbox_key (`Metamessage uid)) 
+      GI_tr.update mbox.trans (get_key mbox.mbox_key (`Metamessage uid)) 
           (Sexp.to_string (sexp_of_mailbox_message_metadata message_metadata)) >>
       parse ~save_message:(fun msg_hash postmark headers content attachments ->
         update_hashes mbox uid msg_hash postmark headers content attachments >>= fun h ->
@@ -677,16 +841,16 @@ module IrminMailbox :
       | `Eof -> return `Eof
       | `NotFound -> return `NotFound
       | `Ok (_,uid) ->
-        IrminIntf_tr.update mbox.trans (get_key mbox.mbox_key (`Metamessage uid))
+        GI_tr.update mbox.trans (get_key mbox.mbox_key (`Metamessage uid))
           (Sexp.to_string (sexp_of_mailbox_message_metadata metadata)) >>= fun () ->
         return `Ok
 
     let read_storage mbox msg_hash contid =
       let key = get_key mbox.mbox_key (`Storage (mbox.user,msg_hash,contid)) in
-      IrminIntf_tr.read_exn mbox.trans key
+      GI_tr.read_exn mbox.trans key
 
     let get_message_metadata mbox uid =
-      IrminIntf_tr.read_exn mbox.trans (get_key mbox.mbox_key (`Metamessage uid)) >>= fun sexp_str ->
+      GI_tr.read_exn mbox.trans (get_key mbox.mbox_key (`Metamessage uid)) >>= fun sexp_str ->
       return (mailbox_message_metadata_of_sexp (Sexp.of_string sexp_str))
 
     let read_lazy_storage mbox lazy_hashes contid =
@@ -740,26 +904,26 @@ module IrminMailbox :
       | `NotFound -> return ()
       | `Ok (seq,uid) ->
         get_hashes mbox1 uid >>= fun h ->
-        IrminIntf_tr.update mbox2.trans (get_key mbox2.mbox_key (`Hashes message_metadata.uid)) 
+        GI_tr.update mbox2.trans (get_key mbox2.mbox_key (`Hashes message_metadata.uid)) 
           (Sexp.to_string (sexp_of_hashes h)) >>
-        IrminIntf_tr.update mbox2.trans (get_key mbox2.mbox_key (`Metamessage message_metadata.uid)) 
+        GI_tr.update mbox2.trans (get_key mbox2.mbox_key (`Metamessage message_metadata.uid)) 
           (Sexp.to_string (sexp_of_mailbox_message_metadata message_metadata)) >>
         update_index_uid mbox2 message_metadata.uid
 
     let remove_storage mbox msg_hash contid =
       let key = get_key mbox.mbox_key (`Storage (mbox.user,msg_hash,contid)) in
-      IrminIntf_tr.remove mbox.trans key
+      GI_tr.remove mbox.trans key
 
     let delete_message mbox position =
       get_uid mbox position >>= function
       | `Ok (_,uid) -> 
         get_hashes mbox uid >>= fun h ->
-        IrminIntf_tr.remove mbox.trans (get_key mbox.mbox_key (`Hashes uid)) >>
+        GI_tr.remove mbox.trans (get_key mbox.mbox_key (`Hashes uid)) >>
         remove_storage mbox h.message h.postmark >>
         remove_storage mbox h.message h.headers >>
         remove_storage mbox h.message h.content >>
-        IrminIntf_tr.remove mbox.trans (get_key mbox.mbox_key (`Metamessage uid)) >>
-        IrminIntf_tr.remove mbox.trans (get_key mbox.mbox_key (`Uid uid)) >>
+        GI_tr.remove mbox.trans (get_key mbox.mbox_key (`Metamessage uid)) >>
+        GI_tr.remove mbox.trans (get_key mbox.mbox_key (`Uid uid)) >>
         let rec delattach i =
           if i < h.attachments then
             remove_storage mbox h.message (string_of_int (3 + i)) >>
@@ -779,7 +943,7 @@ module IrminMailbox :
       |_ -> return ()
 
     let read_subscriptions mbox =
-      IrminIntf_tr.read mbox.trans (get_key mbox.mbox_key `Subscriptions) >>= function
+      GI_tr.read mbox.trans (get_key mbox.mbox_key `Subscriptions) >>= function
       | Some str -> return (list_of_str_sexp str)
       | None -> return []
       
@@ -788,17 +952,18 @@ module IrminMailbox :
       if list_find l (fun i -> i = mbox.mailbox) then 
         return ()
       else
-        IrminIntf_tr.update mbox.trans (get_key mbox.mbox_key `Subscriptions)
+        GI_tr.update mbox.trans (get_key mbox.mbox_key `Subscriptions)
           (str_sexp_of_list (mbox.mailbox :: l))
 
     let unsubscribe mbox =
       read_subscriptions mbox >>= fun l ->
       let l = List.fold_left (fun acc i -> if i = mbox.mailbox then acc else i :: acc) [] l in
-      IrminIntf_tr.update mbox.trans (get_key mbox.mbox_key `Subscriptions)
+      GI_tr.update mbox.trans (get_key mbox.mbox_key `Subscriptions)
         (str_sexp_of_list l)
 
     let list_mailbox mbox key =
-      IrminIntf_tr.list mbox.trans (Key_.add_list mbox.mbox_key (Key_.mailboxes_of_mailbox key))
+      (* sub-mailboxes are located under 'mailboxes' parent *)
+      GI_tr.list mbox.trans (Key_.add_list mbox.mbox_key (Key_.mailboxes_of_mailbox key))
 
     let is_folder mbox key =
       exists_key mbox key >>= function
@@ -847,13 +1012,13 @@ module IrminMailbox :
 
     let list_messages mbox k =
       let list_subtr store k =
-        IrminIntf.list store k >>= fun l ->
+        GI.list store k >>= fun l ->
         return (List.fold_left (fun acc i ->
           acc ^ ":" ^ (List.nth i ((List.length i) - 1))
         ) "" l)
       in
-      IrminIntf.create ~user:mbox.user mbox.config >>= fun store ->
-      IrminIntf.list store k >>= fun l ->
+      GI.create ~user:mbox.user mbox.config >>= fun store ->
+      GI.list store k >>= fun l ->
       Lwt_list.fold_left_s (fun acc i ->
         let k = Key_.t_of_list ((List.concat [k;i])) in
         list_subtr store k >>= fun s ->
@@ -879,7 +1044,11 @@ module IrminMailbox :
 
   end
 
-module UserAccount :
+module GitMailbox = GitMailboxMake(IrminIntf)(IrminIntf_tr)
+
+module GitWorkdirMailbox = GitMailboxMake(GitWorkdirIntf)(GitWorkdirIntf_tr)
+
+module UserAccountMake (GI:GitIntf)(GI_tr:GitIntf_tr) :
   sig
     type t
 
@@ -897,20 +1066,24 @@ module UserAccount :
     (* create new account *)
     let create_account key =
       let (user,config,key) = key in
-      IrminIntf_tr.begin_transaction ~user config key >>= fun view ->
-      IrminIntf_tr.mem view (Key_.t_of_path "subscriptions") >>= fun res ->
+      GI_tr.begin_transaction ~user config key >>= fun view ->
+      GI_tr.mem view (Key_.t_of_path "subscriptions") >>= fun res ->
       if res then
         return `Exists
       else (
-        IrminIntf_tr.update view (Key_.t_of_path "subscriptions") (str_sexp_of_list []) >>
-        IrminIntf_tr.end_transaction view >>
+        GI_tr.update view (Key_.t_of_path "subscriptions") (str_sexp_of_list []) >>
+        GI_tr.end_transaction view >>
         return `Ok
       )
 
     (* remove account *)
     let delete_account key =
       let (user,config,key) = key in
-      IrminIntf.create ~user config >>= fun store ->
-      IrminIntf.remove store key
+      GI.create ~user config >>= fun store ->
+      GI.remove store key
 
   end
+
+module UserAccount = UserAccountMake (IrminIntf) (IrminIntf_tr)
+
+module WorkdirUserAccount = UserAccountMake (GitWorkdirIntf)(GitWorkdirIntf_tr)