+++ lib/commands/server_config.ml
@@ -36,7 +36,7 @@ type imapConfig = {
   key_name : string; (* private key file name, default server.key *)
   pub_name : string; (* public key file name, default server.pub *)
   users_path : string; (* users file path, default datadir/imaplet *)
-  data_store : [`Irmin|`Mailbox|`Maildir]; (* type of storage, only irmin supported so far *)
+  data_store : [`Irmin|`GitWorkdir|`Mailbox|`Maildir]; (* type of storage, only irmin supported so far *)
   encrypt : bool; (* encrypt messages, default true *)
   compress : bool; (* compress messages, but not attachments, default true *)
   user_cert_path : string; (* user's certificate/key location *)
@@ -96,7 +96,7 @@ let validate_config config =
   begin
   match config.data_store with
   | `Irmin -> 
-      let path = Regex.replace ~regx:"%user%.*$" ~tmpl:"" config.irmin_path in
+    let path = Regex.replace ~regx:"%user%.*$" ~tmpl:"" config.irmin_path in
     Utils.exists path Unix.S_DIR >>= fun res ->
     err res "Invalid Irminsule path in"
   | `Mailbox ->
@@ -107,6 +107,10 @@ let validate_config config =
   | `Maildir ->
     Utils.exists config.mail_path Unix.S_DIR >>= fun res ->
     err res "Invalid Maildir path in "
+  | `GitWorkdir ->
+    let path = Regex.replace ~regx:"%user%.*$" ~tmpl:"" config.irmin_path in
+    Utils.exists path Unix.S_DIR >>= fun res ->
+    err res "Invalid Irminsule path in"
   end >>= function
   | `Error err -> return (`Error err)
   | `Ok ->
@@ -160,7 +164,7 @@ let config_of_lines lines =
           let bval n v default = try bool_of_string v with _ -> log n v; default in
           let fval n v default = try float_of_string v with _ -> log n v; default in
           let stval n = function
-            |"irmin"->`Irmin|"mbox"->`Mailbox|"maildir"->`Maildir|_->log n v; `Irmin in
+            |"irmin"->`Irmin|"mbox"->`Mailbox|"maildir"->`Maildir|"gitworkdir"->`GitWorkdir|_->log n v; `Irmin in
           match n with 
           | "imap_name" -> {acc with imap_name = v }
           | "rebuild_irmin" -> {acc with rebuild_irmin = bval n v false}
@@ -233,5 +237,5 @@ let srv_config =
   let config = config_of_lines lines in
   assert ((config.data_store = `Mailbox && config.inbox_path <> "" ||
   config.data_store = `Maildir) && config.mail_path <> "" ||
-  config.data_store = `Irmin);
+  config.data_store = `Irmin || config.data_store = `GitWorkdir);
   config